feat: separate provider from components for easier shadowing
