+++ packages/example/src/pages/components/PageDescription.mdx
@@ -8,7 +8,7 @@ The `<PageDescription>` component is generally used for intro text at the top of
 
 </PageDescription>
 
-Please note that this component is _extremely_ picky about white space, the line break above and below your content is required, and if you even have an extra space on the empty lines above/below it won't render correctly.
+Please note that this component, like all MDX components, is picky about white space, the line break above and below your content is required, and if you even have an extra space on the empty lines above/below it won't render correctly.
 
 ## Example
