+++ packages/gatsby-theme-carbon/src/components/MDXProvider/index.js
@@ -1,0 +1,3 @@
+import MDXProvider from './MDXProvider';
+
+export default MDXProvider;
